+++ src/main/java/BaekJoon/SolvedAC4/BOJ11054_LongestBitonicSequence.java
@@ -1,4 +1,4 @@
-package BaekJoon;
+package BaekJoon.SolvedAC4;
 
 import java.io.BufferedReader;
 import java.io.IOException;
